binance deposit as last step of instant flow
